+++ examples/sph/test/test.cpp
@@ -13,66 +13,66 @@
 
 
 TEST_CASE("test_spiky_kernel_coeff_3D", "[sph]") {
-    double h = 0.789f;
-    CHECK(get_debrun_coeff_3D(h) == Catch::Approx(19.791529914316335).margin(1e-7f));
+    double h = 0.789;
+    CHECK(get_debrun_coeff_3D(h) == Catch::Approx(19.791529914316335).margin(1e-7));
 }
 
 TEST_CASE("test_spiky_kernel", "[sph]") {
-    double h = 1.329f;
-    double r = 0.39881f;
-    REQUIRE(debrun_spiky_kernel(-0.000001f, h) == 0.0f);
-    REQUIRE(debrun_spiky_kernel(1.33f, h) == 0.0f);
-    REQUIRE(debrun_spiky_kernel(1.30f, h) >= 0.0f);
-    REQUIRE(debrun_spiky_kernel(0.1f, h) >= 0.0f);
+    double h = 1.329;
+    double r = 0.39881;
+    REQUIRE(debrun_spiky_kernel(-0.000001, h) == 0.0);
+    REQUIRE(debrun_spiky_kernel(1.33, h) == 0.0);
+    REQUIRE(debrun_spiky_kernel(1.30, h) >= 0.0);
+    REQUIRE(debrun_spiky_kernel(0.1, h) >= 0.0);
 
     double sum = 0;
     int d = 25;
     double maxd = (double) d;
-    double scale = 1.5f;
+    double scale = 1.5;
     // Test that it approximately integrates to 1 by computing a grid
     for (int i = -d; i < d; ++i) {
         for (int j = -d; j < d; ++j) {
             for (int k =  -d; k < d; ++k) {
-                vec3 pos = vec3((double) i * scale / maxd, (double) j * scale / maxd, (double) k * scale / maxd);
+                dvec3 pos = dvec3((double) i * scale / maxd, (double) j * scale / maxd, (double) k * scale / maxd);
                 REQUIRE(pos.x <= scale);
                 REQUIRE(pos.x >= -scale);
                 double dist = glm::length(pos);
-                REQUIRE(dist <= std::sqrt(3.0f) * scale);
+                REQUIRE(dist <= std::sqrt(3.0) * scale);
                 double dx = scale / maxd;
                 double vol = dx * dx * dx;
                 sum += debrun_spiky_kernel(dist, h) * vol;
             }
         }
     }
-    REQUIRE(sum == Catch::Approx(1.0).margin(1e-2f));
+    REQUIRE(sum == Catch::Approx(1.0).margin(1e-2));
 }
 
 TEST_CASE("test_kernel_dwdr", "[sph]") {
-     double h = 1.329f;
-     CHECK(std::fabs(debrun_spiky_kernel_dwdr(0.0f, h)) > 1.0f); // doesn't disappear at origin
+     double h = 1.329;
+     CHECK(std::fabs(debrun_spiky_kernel_dwdr(0.0, h)) > 1.0); // doesn't disappear at origin
 }
 
 TEST_CASE("test_kernel_grad", "[sph]") {
-    double dx = 0.1361f;
-    double dy = 0.9981f;
-    double dz = 0.5012f;
-    vec3 pos = vec3(dx, dy, dz);
-    double h = 1.8f;
+    double dx = 0.1361;
+    double dy = 0.9981;
+    double dz = 0.5012;
+    dvec3 pos = dvec3(dx, dy, dz);
+    double h = 1.8;
     double r = length(pos);
 
     auto grad = debrun_spiky_kernel_grad(pos, h);
 
-    CHECK(std::fabs(grad.x - dx * debrun_spiky_kernel_dwdr(r, h) / r) < 1e-6f);
-    CHECK(std::fabs(grad.y - dy * debrun_spiky_kernel_dwdr(r, h) / r) < 1e-6f);
-    CHECK(std::fabs(grad.z - dz * debrun_spiky_kernel_dwdr(r, h) / r) < 1e-6f);
-    CHECK(grad.x < 0.0f); // we expect the gradient to be pointing downwards
-    CHECK(grad.y < 0.0f);
+    CHECK(std::fabs(grad.x - dx * debrun_spiky_kernel_dwdr(r, h) / r) < 1e-6);
+    CHECK(std::fabs(grad.y - dy * debrun_spiky_kernel_dwdr(r, h) / r) < 1e-6);
+    CHECK(std::fabs(grad.z - dz * debrun_spiky_kernel_dwdr(r, h) / r) < 1e-6);
+    CHECK(grad.x < 0.0); // we expect the gradient to be pointing downwards
+    CHECK(grad.y < 0.0);
 }
 
 TEST_CASE("test_debrun_spiky_kernel_lap", "[sph]") {
-    double h = 1.2f;
-    double dr = 1e-5f;  // small step for finite differences
-    vec3 v1 = vec3(0.9f, 0.9f, 0.9f);
+    double h = 1.2;
+    double dr = 1e-5;  // small step for finite differences
+    dvec3 v1 = dvec3(0.9, 0.9, 0.9);
     double r = length(v1);
 
     // Compute gradient at v1
@@ -81,14 +81,14 @@ TEST_CASE("test_debrun_spiky_kernel_lap", "[sph]") {
     // Finite difference approximation of Laplacian
     double lap_fd = 0.0f;
     for (int i = 0; i < 3; ++i) {
-        vec3 dv = vec3(0.0f);
+        dvec3 dv = dvec3(0.0);
         if (i == 0) dv.x = dr;
         if (i == 1) dv.y = dr;
         if (i == 2) dv.z = dr;
 
         auto grad_forward = debrun_spiky_kernel_grad(v1 + dv, h);
         auto grad_backward = debrun_spiky_kernel_grad(v1 - dv, h);
-        double derivative = (grad_forward[i] - grad_backward[i]) / (2.0f * dr);
+        double derivative = (grad_forward[i] - grad_backward[i]) / (2.0 * dr);
         // derivative of i-th component along its coordinate
         // double derivative = (grad_forwawrd[i] - grad_center[i]) / dr;
         lap_fd += derivative;    
@@ -101,42 +101,42 @@ TEST_CASE("test_debrun_spiky_kernel_lap", "[sph]") {
 }
 
 TEST_CASE("cal_pressure_wcsph behaves correctly", "[sph]") {
-    double rho = 1100.0f;
-    double rho0 = 1000.0f;
-    double c2 = 1500.0f;
-    double gamma = 7.0f;
+    double rho = 1100.0;
+    double rho0 = 1000.0;
+    double c2 = 1500.0;
+    double gamma = 7.0;
 
     double result = cal_pressure_wcsph(rho, rho0, c2, gamma);
     double expected_bweak = c2 * rho0 / gamma;
-    double expected = expected_bweak * (std::pow(rho / rho0, gamma) - 1.0f);
+    double expected = expected_bweak * (std::pow(rho / rho0, gamma) - 1.0);
 
     REQUIRE(std::abs(result - expected) < 1e-6f);
 }
 
 TEST_CASE("cal_rho_ij returns zero outside support radius", "[sph]") {
-    double mass_j = 2.0f;
-    double h = 1.0f;
+    double mass_j = 2.0;
+    double h = 1.0;
 
     // distance outside kernel support
-    double dist_far = 1.1f;
-    REQUIRE(cal_rho_ij(mass_j, dist_far, h) == 0.0f);
+    double dist_far = 1.1;
+    REQUIRE(cal_rho_ij(mass_j, dist_far, h) == 0.0);
 
     // distance inside kernel support
-    double dist_near = 0.5f;
-    REQUIRE(cal_rho_ij(mass_j, dist_near, h) > 0.0f);
+    double dist_near = 0.5;
+    REQUIRE(cal_rho_ij(mass_j, dist_near, h) > 0.0);
 }
 
 TEST_CASE("cal_pressure_force_coefficient computes correctly", "[sph]") {
-    double pi = 2000.0f;
-    double pj = 1500.0f;
-    double rhoi = 1000.0f;
-    double rhoj = 900.0f;
-    double mj = 1.5f;
+    double pi = 2000.0;
+    double pj = 1500.0;
+    double rhoi = 1000.0;
+    double rhoj = 900.0;
+    double mj = 1.5;
 
     double result = cal_pressure_force_coefficient(pi, pj, rhoi, rhoj, mj);
     double expected = ((pi / (rhoi*rhoi)) + (pj / (rhoj*rhoj))) * mj;
 
-    REQUIRE(std::abs(result - expected) < 1e-6f);
+    REQUIRE(std::abs(result - expected) < 1e-6);
 }
 
 
@@ -156,7 +156,7 @@ TEST_CASE("Test that pipeline produces correct density values for d = 0 for firs
     
     size_t nCells = static_cast<uint32_t>(cellData.size());
     size_t printed = 0;
-    double h = 1.5f / (1 << nBitsPerAxis);
+    double h = 1.5 / (1 << nBitsPerAxis);
     std:: cerr << "Checking SPH output with" << nCells << " cells" << std::endl;
 
     for (uint32_t key = 0; key < nCells; ++key) {
@@ -169,7 +169,7 @@ TEST_CASE("Test that pipeline produces correct density values for d = 0 for firs
             continue; // Empty cell
         }
 
-        double dens = 0.0f;
+        double dens = 0.0;
         for (uint32_t pind = start; pind < end; ++pind) {
             uint32_t unsorted_ind = indexData[pind];
             auto particle = inputPos[unsorted_ind];
@@ -206,7 +206,7 @@ TEST_CASE("Test that pipeline produces correct density values for random cell wi
     for (uint32_t p0idx : {0, 27, 35, 109, 111}) {
 
         // choose from output pos so we can match to validation output density
-        glm::vec3 p0 = outputPos[p0idx].data;
+        glm::dvec3 p0 = outputPos[p0idx].data;
         uvec3 ijk = uvec3(
             binPosition(p0.x, out.params.nBits),
             binPosition(p0.y, out.params.nBits),
